Fix bug in overloading for Float64
Ideally we wouldn't have this string typing here, but it works for now.

+++ src/types.rs
@@ -246,7 +246,7 @@ impl TypeSubstitution {
                 for x in xs.iter() {
                     match self.overloads.get(x) {
                         Some(ts) => {
-                            if !ts.contains(&"Integer".into()) {
+                            if !ts.contains(&"Float64".into()) {
                                 return false;
                             }
                         }
